PageContainer, Property Display, Realtime Update Setpup
Page Container (NetworkView):
to pass node properties whenever its been pressed by the user.
Property Display:
OnNodeClick event is added to pass the node's properties to the Property Component with reference to the nodeId. A memoization of the topsort is also made to prevent recomputation  of the layers when re-renders occur; to only be recomputed when the "graphData" is changed.

This commit sets up dynamic updates by preparing for realtime updates of the neural network.